refactor(PhotoLibraryUploader): split class for readability
